Splits DbRepos to a different file
